removed double

+++ src/ofxGaussian.cpp
@@ -17,10 +17,10 @@ float ofxGaussian() {
        }
        while (s >= 1 || s == 0);
 
-       double multiplier = sqrt(-2 * log(s)/s);
+       float multiplier = sqrt(-2 * log(s)/s);
        nextNextGaussian = v2 * multiplier;
        haveNextNextGaussian = true;
 
-       return (float)(v1 * multiplier);
+       return v1 * multiplier;
     }
 }
